Fix duplicate name bug
This commit fixes a bug where _file_opener should refuse to open a file.

In the case where the file hierarchy was 0 levels deep (i.e. the file we
wanted to open was the same as pwd, the string would split incorrectly
due to no slashes being part of the string. Now, the string splits
correctly.

Also, the error message in the second case has been set to indicate a
deliberate action.